Adding current user to a controllers/concerns
